Superiorwalls: Fix bug preventing the Bottom Navigation bar to detect clicks
This commit removes `setupContentBottomOffset()` from `onAttach()` of BaseFramesFragment.kt because view and getView() are both null at this point. Additionally, we're getting the measured height of `bottomNavigation` only if one of the views on which the padding should be applied isn't null. This prevents the initialization of `bottomNavigation` too early.

Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>
Change-Id: I6f468445e0a21e21fd58695d40e5f18c1e40b688